% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diesel.R
\docType{data}
\name{diesel}
\alias{diesel}
\title{Data of Diesel prices by Year and Month}
\format{
A data.frame with 243 observations of 3 variables
\describe{
\item{Jahr}{dbl Jahr, in dem der Preis beobachtet wurde}
\item{Month}{dbl Monat, in dem der Preis beobachtet wurde, von 1-12 (Januar bis Dezember)}
\item{Preis in US-Dollar}{dbl Preis in US-Dollar}
}
}
\usage{
diesel
}
\description{
Data of Diesel prices by Year and Month
}
\keyword{datasets}
