% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.dhge-package.R
\docType{package}
\name{stats.dhge-package}
\alias{stats.dhge}
\alias{stats.dhge-package}
\title{stats.dhge: Testpaket für Statistik}
\description{
JOAR.
}
\author{
\strong{Maintainer}: Daniel Hoppe \email{Daniel.Hoppe@dhge.de}

}
\keyword{internal}
